+++ src/main/java/net/geforcemods/securitycraft/network/client/PlayAlarmSound.java
@@ -5,6 +5,7 @@ import net.geforcemods.securitycraft.SecurityCraft;
 import net.geforcemods.securitycraft.blockentities.AlarmBlockEntity;
 import net.minecraft.client.Minecraft;
 import net.minecraft.tileentity.TileEntity;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvent;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -12,19 +13,18 @@ import net.minecraftforge.fml.common.network.ByteBufUtils;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
-import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
 public class PlayAlarmSound implements IMessage {
 	private BlockPos bePos;
-	private SoundEvent sound;
+	private ResourceLocation sound;
 	private int soundX, soundY, soundZ;
 	private float volume, pitch;
 
 	public PlayAlarmSound() {}
 
-	public PlayAlarmSound(BlockPos bePos, SoundEvent sound, float volume, float pitch) {
+	public PlayAlarmSound(BlockPos bePos, ResourceLocation sound, float volume, float pitch) {
 		this.bePos = bePos;
 		this.sound = sound;
 		this.soundX = (int) (bePos.getX() * 8.0F);
@@ -37,7 +37,7 @@ public class PlayAlarmSound implements IMessage {
 	@Override
 	public void toBytes(ByteBuf buf) {
 		buf.writeLong(bePos.toLong());
-		ByteBufUtils.writeRegistryEntry(buf, sound);
+		ByteBufUtils.writeUTF8String(buf, sound.toString());
 		buf.writeInt(soundX);
 		buf.writeInt(soundY);
 		buf.writeInt(soundZ);
@@ -48,7 +48,7 @@ public class PlayAlarmSound implements IMessage {
 	@Override
 	public void fromBytes(ByteBuf buf) {
 		bePos = BlockPos.fromLong(buf.readLong());
-		sound = ByteBufUtils.readRegistryEntry(buf, ForgeRegistries.SOUND_EVENTS);
+		sound = new ResourceLocation(ByteBufUtils.readUTF8String(buf));
 		soundX = buf.readInt();
 		soundY = buf.readInt();
 		soundZ = buf.readInt();
@@ -78,7 +78,7 @@ public class PlayAlarmSound implements IMessage {
 
 				if (te instanceof AlarmBlockEntity) {
 					((AlarmBlockEntity) te).setPowered(true);
-					((AlarmBlockEntity) te).playSound(message.getX(), message.getY(), message.getZ(), message.sound, message.volume, message.pitch);
+					((AlarmBlockEntity) te).playSound(message.getX(), message.getY(), message.getZ(), new SoundEvent(message.sound), message.volume, message.pitch);
 				}
 			});