+++ Changelog.md
@@ -67,6 +67,7 @@
 - Fix: Shadow around the Bouncing Betty, IMS, Mine, and Portable Radar
 - Fix: The Bouncing Betty, IMS, Mine, and Portable Radar cause suffocation
 - Fix: The collision of Block Pocket Wall blocks gets desynchronized when activating a block pocket
+- Fix: The alarm can't play sounds added by resource packs
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
 - Misc.: Small optimizations to certain network and client performance